backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Still needed for Alembic migrations
geoalchemy2==0.14.3
alembic==1.13.1
greenlet>=3.0.1  # Required for async SQLAlchemy (flexible version for playwright compatibility)

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP client for data ingestion
httpx==0.26.0
aiohttp==3.9.1

# Web scraping (for SAGI data)
playwright==1.40.0
beautifulsoup4==4.12.2

# Data processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2  # Excel file processing
shapely==2.0.2
geojson==3.1.0

# Scheduling
apscheduler==3.10.4

# Google Generative AI (Gemini)
google-generativeai==0.8.3

# Twilio (WhatsApp, SMS, Voice)
twilio==9.0.0

# PDF and QR Code generation
reportlab==4.0.8
qrcode==7.4.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
orjson==3.9.10
tenacity==8.2.3
structlog==24.1.0
prometheus-client==0.19.0

# MCP (Model Context Protocol) support
# Note: MCPs run externally via npx/uvx, these are Python utilities
chromadb==0.4.22  # Vector database for RAG
pillow==10.2.0  # Image processing for OCR
pytesseract==0.3.10  # Backup OCR (if MCP unavailable)

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
aiosqlite==0.19.0  # For async SQLite in tests
black==23.12.1
ruff==0.1.11
mypy==1.8.0
types-redis==4.6.0.11
